pyproject.toml
[project]
name = "test-graphdb"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "neo4j>=6.1.0",
    "neo4j-graphrag[openai]>=1.13.0",
    "python-dotenv>=1.2.1",
]
